src/requirements.txt
absl-py==2.1.0
accelerate==1.4.0
Adafruit-Blinka==8.22.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-framebuf==1.6.3
adafruit-circuitpython-requests==2.0.1
adafruit-circuitpython-ssd1306==2.12.14
adafruit-circuitpython-typing==1.9.4
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.52.0
Adafruit-PureIO==1.1.11
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==3.7.1
appdirs==1.4.4
arrow==1.3.0
async-timeout==4.0.3
attrs==25.1.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.2
beautifulsoup4==4.13.3
blinker==1.9.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.1
cachetools==5.3.1
caldav==1.3.9
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clean-fid==0.1.35
click==8.1.8
clip-anytorch==2.6.0
cloudpathlib==0.20.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==0.29.28
dataclasses-json==0.6.7
dateparser==1.1.8
dctorch==0.1.2
decorator==5.2.1
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
encodec==0.1.1
epitran==1.17
espeak_phonemizer==1.3.1
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.3.1
g2pkk==0.1.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
grpcio==1.70.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fa==2.0.1
gruut_lang_fr==2.0.2
gruut_lang_it==2.0.1
gruut_lang_nl==2.0.2
gruut_lang_ru==2.0.1
gruut_lang_sw==2.0.1
gTTS==2.5.4
h11==0.14.0
h2==4.2.0
hangul-romanize==0.1.0
hazm==0.7.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
humanfriendly==10.0
Hypercorn==0.17.3
hyperframe==6.1.0
icalendar==6.1.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
inflect==5.6.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonlines==1.2.0
jsonmerge==1.9.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
k-diffusion==0.1.1.post1
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.43
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langcodes==3.5.0
langgraph==0.3.5
langgraph-checkpoint==2.0.18
langgraph-prebuilt==0.1.2
langgraph-sdk==0.1.55
langsmith==0.3.13
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.2.post1
libwapiti==0.2.1
litellm==1.41.24
llvmlite==0.44.0
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
mecab-python3==1.0.5
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
munkres==1.1.4
murmurhash==1.0.12
mutagen==1.47.0
mycroft_mimic3_tts==0.2.4
mypy-extensions==1.0.0
networkx==2.8.8
ninja==1.11.1
nltk==3.3
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.65.4
orjson==3.10.15
packaging==24.2
pandas==2.2.3
panphon==0.21.2
phonemes2ids==1.2.2
phue==1.1
picovoice==3.0.4
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
priority==2.0.0
propcache==0.3.0
protobuf==3.19.6
psutil==7.0.0
pvporcupine==3.0.3
pvrecorder==1.2.7
pvrhino==3.0.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
pyftdi==0.55.0
pygame==2.6.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
pypinyin==0.53.0
pysbd==0.3.4
pyserial==3.5
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-vlc==3.0.21203
pyttsx3==2.90
pytube==15.0.0
pytz==2025.1
pyusb==1.2.1
PyYAML==6.0.2
Quart==0.20.0
quart-cors==0.8.0
recurring-ical-events==3.5.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
RPi.GPIO==0.7.1
rsa==4.9
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
semantic-router==0.0.72
sentry-sdk==2.22.0
setproctitle==1.3.5
shellingham==1.5.4
simpleaudio==1.0.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
spidev==3.6
spotipy==2.23.0
SQLAlchemy==2.0.38
srsly==2.5.1
starlette==0.46.0
SudachiDict-core==20250129
SudachiPy==0.6.10
swagger-ui-py==21.12.8
sympy==1.13.1
taskgroup==0.2.2
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchdiffeq==0.2.5
torchsde==0.2.6
torchvision==0.21.0
tqdm==4.67.1
trainer==0.0.20
trampoline==0.1.2
transformers==4.49.0
typeguard==4.4.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
umap-learn==0.5.7
unicodecsv==0.14.1
Unidecode==1.3.8
unidic-lite==1.0.8
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.23.2
vobject==0.9.9
wandb==0.19.8
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
x-wr-timezone==2.0.1
xdgenvpy==2.4.0
yarl==1.18.3
yt-dlp==2025.3.31
zipp==3.21.0
zstandard==0.23.0
